2003-03-29  Christian Stimming  <stimming@tuhh.de>
        * src/engine/gnc-session.c: OpenBSD fix by Todd T. Fries
        <todd@flare.fries.net>

        * src/engine/Transaction.c: OpenBSD fix by Todd T. Fries
        <todd@flare.fries.net>: time_t is not long everywhere.


git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@8104 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/engine/Transaction.c
@@ -2492,7 +2492,7 @@ xaccTransSetDateInternal(Transaction *trans, int which, time_t secs,
     check_open(trans);
 
     PINFO ("addr=%p set %d date to %lu %li %s",
-           trans, which, secs, nsecs, ctime (&secs));
+	   trans, which, (long unsigned int)secs, nsecs, ctime (&secs));
     
     dadate = ((which == TDATE_POSTED)
               ? &trans->date_posted
